Настроил stor. Выбранный уровень сохраняется в stor.
