+++ src/lib/LEVELS.js
@@ -1,0 +1,5 @@
+export const LEVELS = [
+    { id: 'radio1', value: '1', label: '1' },
+    { id: 'radio2', value: '2', label: '2' },
+    { id: 'radio3', value: '3', label: '3' },
+];
