--- src/lib/cardArr.js
@@ -1,3 +1,0 @@
-export const cardArr = [
-    
-]
