+++ src/pages/main/mainPage.jsx
@@ -1,7 +1,16 @@
+import { useDispatch } from 'react-redux';
 import { Difficulty } from '../../components/Difficulty/Difficulty';
 import * as S from './style/mainPage.S'
+import { useEffect } from 'react';
+import { setLevel } from '../../store/slices/slices';
 
 export const MainPage = () => {
+
+    const dispatch = useDispatch()
+
+    useEffect(() => {
+        dispatch(setLevel(null))
+    },[])
 
     return (
