+++ src/store/store.js
@@ -1,0 +1,8 @@
+import { configureStore } from "@reduxjs/toolkit";
+import gameReducer from "./slices/slices";
+
+export const store = configureStore({
+    reducer: {
+        game: gameReducer,
+    },
+});
